nas/tensorflow/requirements.txt
absl-py==0.1.10
alabaster==0.7.10
appdirs==1.4.3
astroid==1.5.2
Babel==2.4.0
backports-abc==0.5
backports.functools-lru-cache==1.3
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
backports.weakref==1.0.post1
bleach==1.5.0
certifi==2017.4.17
chardet==3.0.2
configparser==3.5.0
cycler==0.10.0
Cython==0.25.2
decorator==4.0.11
docutils==0.13.1
easydict==1.7
entrypoints==0.2.2
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.2.0
grpc==0.3.post19
grpcio==1.8.4
html5lib==0.9999999
imagesize==0.7.1
ipykernel==4.6.1
ipython==5.3.0
ipython-genutils==0.2.0
isort==4.2.5
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-client==5.0.1
jupyter-core==4.3.0
lazy-object-proxy==1.2.2
Mako==1.0.6
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.0.0
mccabe==0.6.1
mistune==0.7.4
mock==2.0.0
msgpack-python==0.5.1
nbconvert==5.1.1
nbformat==4.3.0
networkx==1.11
numpy==1.13.3
numpydoc==0.6.0
olefile==0.44
packaging==16.8
pandocfilters==1.4.1
pathlib2==2.2.1
pbr==3.1.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
prompt-toolkit==1.0.14
protobuf==3.5.0.post1
psutil==5.2.2
ptyprocess==0.5.1
py==1.4.34
pycosat==0.6.1
pyflakes==1.5.0
Pygments==2.2.0
pylint==1.7.1
pymongo==3.6.1
pyparsing==2.2.0
pytesseract==0.1.7
pytest==3.1.2
python-dateutil==2.6.0
pytools==2017.3
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.0
QtPy==1.2.1
requests==2.13.0
rope==0.10.5
ruamel.ordereddict==0.4.9
ruamel.yaml==0.14.11
scandir==1.5
scikit-image==0.13.0
scikit-learn==0.18.1
scipy==0.19.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
snowballstemmer==1.2.1
spearmint==0.1
Sphinx==1.5.5
spyder==2.3.9
subprocess32==3.2.7
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0
testpath==0.3
Theano==0.8.0
tornado==4.5.1
traitlets==4.3.2
wcwidth==0.1.7
weave==0.16.0
webencodings==0.5.1
Werkzeug==0.12.2
wrapt==1.10.10
